@@ -34,7 +34,7 @@
     "build": "trash dist && yarn build:main",
     "build:main": "tsc -p tsconfig.build.json",
     "lint": "tslint --project tsconfig.json --config tslint.json",
-    "unit": "jest --forceExit",
+    "unit": "jest --detectOpenHandles",
     "test": "yarn lint && yarn unit",
     "test:integration": "yarn lint && jest --config=jest-integration.config.js",
     "watch": "jest --watch",
@@ -43,8 +43,8 @@
     "send-coverage": "codecov",
     "docs": "yarn docs:html && open-cli docs/index.html",
     "docs:test": "yarn docs:html",
-    "docs:html": "typedoc src/index.ts --excludePrivate --mode file --theme minimal --out docs",
-    "docs:json": "typedoc --mode file --json docs/typedoc.json src/index.ts",
+    "docs:html": "typedoc src/index.ts --excludePrivate --theme minimal --out docs",
+    "docs:json": "typedoc --json docs/typedoc.json src/index.ts",
     "docs:publish": "yarn docs:html && gh-pages -d docs",
     "changelog": "standard-version",
     "release": "yarn reset && yarn test && yarn docs:publish && yarn changelog",
@@ -86,21 +86,22 @@
     "/LICENSE"
   ],
   "devDependencies": {
-    "@types/jest": "^25.2.3",
+    "@types/jest": "^26.0.22",
     "@types/node": "^12.12.3",
-    "codecov": "^3.6.5",
-    "gh-pages": "^3.0.0",
-    "jest": "^24.9.0",
+    "codecov": "^3.8.1",
+    "gh-pages": "^3.1.0",
+    "jest": "^26.6.3",
+    "mockdate": "^3.0.5",
     "node-license-validator": "^1.3.0",
     "npm-scripts-info": "^0.3.9",
     "open-cli": "^6.0.1",
-    "standard-version": "^8.0.0",
-    "trash-cli": "^3.0.0",
-    "ts-jest": "^24.0.2",
-    "tslint": "^6.1.2",
+    "standard-version": "^9.2.0",
+    "trash-cli": "^4.0.0",
+    "ts-jest": "^26.5.4",
+    "tslint": "^6.1.3",
     "tslint-config-standard": "^9.0.0",
-    "typedoc": "^0.16.8",
-    "typescript": "~3.6.5"
+    "typedoc": "^0.20.35",
+    "typescript": "~4.1.5"
   },
   "keywords": [
     "mos",
@@ -117,9 +118,9 @@
     "production"
   ],
   "dependencies": {
-    "iconv-lite": "^0.5.0",
-    "moment": "^2.26.0",
-    "tslib": "^1.13.0",
+    "iconv-lite": "^0.6.2",
+    "moment": "^2.29.1",
+    "tslib": "^2.2.0",
     "xml-js": "^1.6.11",
     "xmlbuilder": "^15.1.1"
   },
