@@ -1,8 +1,10 @@
 module.exports = {
 	globals: {
 		'ts-jest': {
-			tsConfig: 'tsconfig.json',
-			diagnostics: false
+			tsconfig: 'tsconfig.json',
+			diagnostics: {
+				ignoreCodes: []
+			}
 		}
 	},
 	moduleFileExtensions: [
@@ -17,7 +19,7 @@ module.exports = {
 	],
 	testPathIgnorePatterns: [
 		'integrationTests'
-	],	
+	],
 	testEnvironment: 'node',
 	coverageThreshold: {
 		global: {
