@@ -2,8 +2,8 @@
 export * from './MosConnection'
 export * from './api'
 export {
-	IConnectionConfig,
-	ConnectionConfig
+	ConnectionConfig,
+	IProfiles
 } from './config/connectionConfig'
 
 export { MosString128 } from './dataTypes/mosString128'
@@ -11,3 +11,7 @@ export { MosDuration } from './dataTypes/mosDuration'
 export { IMOSScope, IMOSExternalMetaData } from './dataTypes/mosExternalMetaData'
 export { MosTime } from './dataTypes/mosTime'
 export { MosDevice } from './MosDevice'
+
+export {
+	MosItemReplaceOptions
+} from './mosModel'
