@@ -1,6 +1,6 @@
 import * as XMLBuilder from 'xmlbuilder'
 import { MosTime } from '../../dataTypes/mosTime'
-import { MosMessage } from '../MosMessage'
+import { MosMessage, PortType } from '../MosMessage'
 import { addTextElement } from '../../utils/Utils'
 
 export class HeartBeat extends MosMessage {
@@ -8,14 +8,15 @@ export class HeartBeat extends MosMessage {
 	time: MosTime
 
   /** */
-	constructor (time: MosTime = new MosTime()) {
-		super()
+	constructor (port: PortType, time: MosTime = new MosTime()) {
+		super(port)
 		this.time = time
 	}
 
   /** */
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
-		let messageBlock = addTextElement(XMLBuilder.create('heartbeat'), 'time', {}, this.time)
-		return messageBlock
+		const heartbeat = XMLBuilder.create('heartbeat')
+		addTextElement(heartbeat, 'time', this.time)
+		return heartbeat
 	}
 }
