@@ -1,7 +1,7 @@
 import { MosMessage } from '../MosMessage'
 import * as XMLBuilder from 'xmlbuilder'
 import { IMOSObject } from '../../api'
-import { Parser } from '../Parser'
+import { XMLMosObject } from './xmlConversion'
 
 export class MosObj extends MosMessage {
 
@@ -9,13 +9,15 @@ export class MosObj extends MosMessage {
 
   /** */
 	constructor (obj: IMOSObject) {
-		super()
+		super('lower')
 		this.obj = obj
-		this.port = 'lower'
 	}
 
   /** */
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
-		return Parser.mosObj2xml(this.obj)
+		let xmlMosObj = XMLBuilder.create('mosObj')
+		XMLMosObject.toXML(xmlMosObj, this.obj)
+
+		return xmlMosObj
 	}
 }
