@@ -8,15 +8,14 @@ export class ReqMosObj extends MosMessage {
 	private objId: MosString128
   /** */
 	constructor (objId: MosString128) {
-		super()
-		this.port = 'lower'
+		super('lower')
 		this.objId = objId
 	}
 
   /** */
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
 		let root = XMLBuilder.create('mosReqObj')
-		addTextElement(root, 'objID', {}, this.objId)
+		addTextElement(root, 'objID', this.objId)
 		return root
 	}
 }
