@@ -4,8 +4,8 @@ import {
 	IMOSRunningOrder,
 	IMOSROStory
 } from '../../api'
-import { Parser } from '../Parser'
 import { addTextElement } from '../../utils/Utils'
+import { XMLROStory } from './xmlConversion'
 
 export class ROList extends MosMessage {
 
@@ -14,19 +14,18 @@ export class ROList extends MosMessage {
 
   /** */
 	constructor () {
-		super()
+		super('upper')
 	}
 
   /** */
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
 		let root = XMLBuilder.create('roList')
 
-		addTextElement(root, 'roID', {}, this.RO.ID)
-		addTextElement(root, 'roSlug', {}, this.RO.Slug)
+		addTextElement(root, 'roID', this.RO.ID)
+		addTextElement(root, 'roSlug', this.RO.Slug)
 
 		this.RO.Stories.forEach((story: IMOSROStory) => {
-			let xmlStory = Parser.story2xml(story)
-			root.importDocument(xmlStory)
+			XMLROStory.toXML(root, story)
 		})
 
 		return root
