@@ -24,10 +24,9 @@ export class ROElementStat extends MosMessage {
 	private time: MosTime
   /** */
 	constructor (options: ROElementStatOptions) {
-		super()
+		super('upper')
 		this.options = options
 		this.time = new MosTime()
-		this.port = 'upper'
 	}
 
   /** */
@@ -35,13 +34,13 @@ export class ROElementStat extends MosMessage {
 		let root = XMLBuilder.create('roElementStat')
 		root.attribute('element', this.options.type.toString())
 
-		addTextElement(root, 'roID', {}, this.options.roId)
-		if (this.options.storyId) 		addTextElement(root, 'storyID', {}, this.options.storyId)
-		if (this.options.itemId) 		addTextElement(root, 'itemID', {}, this.options.itemId)
-		if (this.options.objId) 		addTextElement(root, 'objID', {}, this.options.objId)
-		if (this.options.itemChannel) 	addTextElement(root, 'itemChannel', {}, this.options.itemChannel)
-		addTextElement(root, 'status', {}, this.options.status)
-		addTextElement(root, 'time', {}, this.time)
+		addTextElement(root, 'roID', this.options.roId)
+		if (this.options.storyId) 		addTextElement(root, 'storyID', this.options.storyId)
+		if (this.options.itemId) 		addTextElement(root, 'itemID', this.options.itemId)
+		if (this.options.objId) 		addTextElement(root, 'objID', this.options.objId)
+		if (this.options.itemChannel) 	addTextElement(root, 'itemChannel', this.options.itemChannel)
+		addTextElement(root, 'status', this.options.status)
+		addTextElement(root, 'time', this.time)
 		return root
 	}
 }
