@@ -1,9 +1,9 @@
 import { IMOSItem } from '../../api'
 import { MosMessage } from '../MosMessage'
 import * as XMLBuilder from 'xmlbuilder'
-import { Parser } from '../Parser'
 import { MosString128 } from '../../dataTypes/mosString128'
 import { addTextElement } from '../../utils/Utils'
+import { XMLMosItem } from '../profile2/xmlConversion'
 
 export interface MosItemReplaceOptions {
 	roID: MosString128
@@ -15,18 +15,18 @@ export class MosItemReplace extends MosMessage {
 	private options: MosItemReplaceOptions
 
 	constructor (options: MosItemReplaceOptions) {
-		super()
+		super('upper')
 		this.options = options
-		this.port = 'upper'
 	}
 
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
 		const item = this.options.item
 		const root = XMLBuilder.create('mosItemReplace')
 
-		addTextElement(root, 'roID', {}, this.options.roID)
-		addTextElement(root, 'storyID', {}, this.options.storyID)
-		root.importDocument(Parser.item2xml(item))
+		addTextElement(root, 'roID', this.options.roID)
+		addTextElement(root, 'storyID', this.options.storyID)
+
+		XMLMosItem.toXML(root, item)
 
 		return root
 	}
