@@ -1,21 +1,21 @@
 import { IMOSObject } from '../../api'
 import { MosMessage } from '../MosMessage'
 import * as XMLBuilder from 'xmlbuilder'
-import { Parser } from '../Parser'
+import { XMLMosObject } from '../profile1/xmlConversion'
 
+// http://mosprotocol.com/wp-content/MOS-Protocol-Documents/MOS-Protocol-2.8.4-Current.htm#mosObjCreate
 export class MosObjCreate extends MosMessage {
 	private object: IMOSObject
 
 	constructor (object: IMOSObject) {
-		super()
+		super('lower')
 		this.object = object
-		this.port = 'lower'
 	}
 
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
 		let xml = XMLBuilder.create('mosObjCreate')
 
-		Parser.attachMosObj2xml(this.object, xml)
+		XMLMosObject.toXML(xml, this.object)
 
 		return xml
 	}
